requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.in -o requirements.txt
annotated-types==0.7.0
    # via pydantic
beautifulsoup4==4.14.3
    # via
    #   markdownify
    #   python-jobspy
certifi==2025.11.12
    # via requests
charset-normalizer==3.4.4
    # via requests
colorama==0.4.6
    # via loguru
idna==3.11
    # via requests
loguru==0.7.3
    # via -r requirements.in
markdownify==0.13.1
    # via python-jobspy
numpy==1.26.3
    # via
    #   pandas
    #   python-jobspy
pandas==2.3.3
    # via
    #   -r requirements.in
    #   python-jobspy
pydantic==2.12.5
    # via python-jobspy
pydantic-core==2.41.5
    # via pydantic
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.1
    # via -r requirements.in
python-jobspy==1.1.82
    # via -r requirements.in
pytz==2025.2
    # via pandas
regex==2024.11.6
    # via python-jobspy
requests==2.32.5
    # via python-jobspy
six==1.17.0
    # via
    #   markdownify
    #   python-dateutil
soupsieve==2.8.1
    # via beautifulsoup4
tls-client==1.0.1
    # via python-jobspy
typing-extensions==4.15.0
    # via
    #   beautifulsoup4
    #   pydantic
    #   pydantic-core
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pandas
urllib3==2.6.2
    # via requests
win32-setctime==1.2.0
    # via loguru
